src/test/java/speque/springframework/spring6restmvc/service/BeerServiceCSVImplTest.java
package speque.springframework.spring6restmvc.service;

import org.junit.jupiter.api.Test;
import org.springframework.util.ResourceUtils;
import speque.springframework.spring6restmvc.model.BeerCSVRecord;

import java.io.File;
import java.io.FileNotFoundException;
import java.util.List;

import static org.assertj.core.api.Assertions.assertThat;
import static org.junit.jupiter.api.Assertions.*;

class BeerServiceCSVImplTest {
    BeerServiceCSV beerCsvService = new BeerServiceCSVImpl();

    @Test
    void convertCSV() throws FileNotFoundException {
        File file = ResourceUtils.getFile("classpath:csvdata/beers.csv");
        List<BeerCSVRecord> recs = beerCsvService.convertCSV(file);
        System.out.println(recs.size());
        assertThat(recs.size()).isGreaterThan(0);
    }
}
